Move Plugin header to interface folder
