+++ tools/cgcollector2/include/cgcollector2/collector/FileInfoCollector.h
@@ -5,8 +5,8 @@
 */
 #ifndef CGCOLLECTOR2_FILEINFOCOLLECTOR_H
 #define CGCOLLECTOR2_FILEINFOCOLLECTOR_H
+#include "cgcollector2/interface/Plugin.h"
 #include "metadata/FileInfoMD.h"
-#include "cgcollector2/Plugin.h"
 
 #include <clang/AST/ASTContext.h>
 #include <clang/Basic/SourceManager.h>
@@ -18,7 +18,7 @@
  * The generated metadata is metacg compatible and defined above
  */
 
-struct FileInfoCollector : Plugin {
+struct FileInfoCollector : cgcollector2::Plugin {
 
   virtual std::unique_ptr<metacg::MetaData> computeForDecl(const clang::FunctionDecl* const functionDecl) {
     std::unique_ptr<FileInfoMetadata> result = std::make_unique<FileInfoMetadata>();
