+++ tools/cgcollector2/include/cgcollector2/collector/LoopDepthCollector.h
@@ -9,9 +9,9 @@
 
 #include "metacg/metadata/LoopMD.h"
 #include "cgcollector2/MetaDataFunctions.h"
-#include "cgcollector2/Plugin.h"
+#include "cgcollector2/interface/Plugin.h"
 
-struct LoopDepthCollector : public Plugin {
+struct LoopDepthCollector : public cgcollector2::Plugin {
   virtual  std::unique_ptr<metacg::MetaData> computeForDecl(clang::FunctionDecl const* const decl) {
     std::unique_ptr<metacg::LoopDepthMD> result = std::make_unique<metacg::LoopDepthMD>();
     result->loopDepth = getLoopDepthInStmt(decl->getBody());
