+++ tools/cgcollector2/src/CallGraphCollectionAction.cpp
@@ -7,7 +7,7 @@
 #include "metacg/MCGBaseInfo.h"
 #include "cgcollector2/CallGraphCollectionAction.h"
 #include "cgcollector2/CallGraphNodeGenerator.h"
-#include "cgcollector2/Plugin.h"
+#include "cgcollector2/interface/Plugin.h"
 #include "cgcollector2/SharedDefs.h"
 #include "cgcollector2/metadata/Internal/ASTNodeMetadata.h"
 #include "cgcollector2/metadata/Internal/AllAliasMetadata.h"
